implement tests/tseries/offsets dir per request
