+++ pandas/tests/tseries/offsets/test_offsets.py
